trydjango19/requirements.txt
Django==1.10.3
django-crispy-forms==1.6.1
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.3
Markdown==2.6.7
markdown2==2.3.2
Pillow==3.4.2
